Fix errors with call signatures in ConcreteElasticASRModel ref #22
This model was not actually being called because of changes in the
function call signatures. As a result, ASR_dependent_E didn’t have any
effect. Fix that, and mark these override so this won’t happen again.

This change does not actually break any tests because the tests as
they are now only have a trivial amount of ASR reaction.