Automate and run complete set of NEML regression tests ref #197
Instead of having copies of the xml and csv files, this adds a script
that points to the versions in NEML using NEML_DIR and populates the
'tests' file.  This gets run when 'make' is run.

This ensures that we will always be running the full set of NEML
tests when we update NEML.