Fix bug in temperature units in SpecifiedVaporPressureBC closes #223
This BC has an equation that computes the relative humidity to be
prescribed using a Dirichlet BC from a formula that is dependent on the
temperature, which should be converted from C to K.  However, the code
only did that conversion in one of the places where the temperature was
used. This corrects that error.