Fix remaining tests that diff due to recent libmesh update closes #259
One of the tests (concrete_ASR_swelling) was simply missed in the
previous minor updates to a few gold files because it was heavy.

The other test was re-golded to pass on linux, but the test failed a
time step on linux, so the time steps didn't match up on mac. I added
the use of a full tangent preconditioner, which makes it converge more
reliably so we shouldn't get steps failing on some machines.