Prevent name clashes for multiple EqualValueEmbeddedConstraintActions ref #299
The names of Constraint objects created by EqualValueEmbeddedConstraintAction
were not connected to variable names, which meant that if more than one instance
of that block were created to allow use with another variable, there would be
a name clash. Fix that by including the variable name in the name of the Constraint.