Update concrete assessment cases after removing deprecated syntax ref #318
A recent change removed usage of deprecated syntax in these tests.
This resulted in an unexpected change in the model behavior because
of a memory access bug that was only triggered by the old syntax.
This updates the gold results for the Wald block tests.

The newer results show greater expansion than before, but the difference
is minor, and in many cases gives better agreement with experimental
results.