Fix valgrind failure closes #354
This new test failed valgrind due to a timeout.
Switched the test from a 4x4x4 cube to a 2x2x2 cube, which drops
the run time with opt from about 6 s to 1 s, which should
be plenty fast to avoid the valgrind timeout. Results are identical
to the prior ones because the solution had a uniform field.